requirements_Version4.txt
flask
flask_sqlalchemy
pyyaml
feedparser
requests